+++ src/gui/src/lib/SignalRecorder.cpp
@@ -15,8 +15,8 @@ SignalRecorder::SignalRecorder(std::size_t size)
 }
 
 std::size_t SignalRecorder::size() const { return time_values.size(); }
-const float* SignalRecorder::time_series() const { return time_values.data(); }
-const float* SignalRecorder::state_series() const { return state_values.data(); }
+const double* SignalRecorder::time_series() const { return time_values.data(); }
+const double* SignalRecorder::state_series() const { return state_values.data(); }
 
 void SignalRecorder::add(Edge edge)
 {
@@ -24,9 +24,9 @@ void SignalRecorder::add(Edge edge)
 
     std::copy(state_values.begin() + 2, state_values.end(), state_values.begin());
     std::copy(time_values.begin() + 2, time_values.end(), time_values.begin());
-    state_values[state_values.size() - 2] = (edge == Edge::Front::RISING) ? 0.f : 1.f;
+    state_values[state_values.size() - 2] = (edge == Edge::Front::RISING) ? 0. : 1.;
     time_values[time_values.size() - 2] = time;
 
-    state_values[state_values.size() - 1] = (edge == Edge::Front::RISING) ? 1.f : 0.f;
+    state_values[state_values.size() - 1] = (edge == Edge::Front::RISING) ? 1. : 0.;
     time_values[time_values.size() - 1] = time + 1;
 }
